refactor(semantic): extract protocol variance checking to ProtocolChecker
