+++ src/semantic/pipeline/mod.rs
@@ -7,5 +7,6 @@ mod type_resolver;
 pub(in crate::semantic) use signature_inference_pass::SignatureInferencePass;
 pub(in crate::semantic) use symbol_collector::SymbolCollector;
 pub(in crate::semantic) use type_checker::TypeChecker;
+pub(in crate::semantic) use type_checker::ProtocolChecker;
 pub(in crate::semantic) use type_constraint_engine::TypeConstraintEngine;
 pub(in crate::semantic) use type_resolver::TypeResolver;
