+++ src/semantic/pipeline/type_checker/mod.rs
@@ -13,13 +13,6 @@ use super::super::{
 };
 use super::{SymbolCollector, TypeConstraintEngine, TypeResolver};
 
-struct InheritedProtocolMethod {
-    name: String,
-    type_id: TypeId,
-    param_types: Vec<SemanticType>,
-    return_type: SemanticType,
-}
-
 mod binary_expr_checker;
 mod block_expr_checker;
 mod builtin_call_expr_checker;
@@ -31,9 +24,12 @@ mod literal_expr_checker;
 mod member_access_expr_checker;
 mod method_call_expr_checker;
 mod new_expr_checker;
+mod protocol_checker;
 mod unary_expr_checker;
 mod variable_expr_checker;
 mod while_expr_checker;
+
+pub(in crate::semantic) use protocol_checker::ProtocolChecker;
 
 pub(in crate::semantic) struct TypeChecker<'a> {
     pub(in crate::semantic) analyzer: &'a mut SemanticAnalyzer,
@@ -323,7 +319,7 @@ impl<'a> TypeChecker<'a> {
                 return;
             };
             for parent_method in
-                self.collect_inherited_protocol_methods(protocol_id, parent_id)
+                ProtocolChecker::collect_inherited_protocol_methods(self.analyzer, protocol_id, parent_id)
             {
                 let key = SymbolCollector::method_symbol_key(protocol_id, &parent_method.name);
                 if !self.analyzer.functions.contains_key(&key) {
@@ -348,119 +344,13 @@ impl<'a> TypeChecker<'a> {
         }
     }
 
-    fn collect_inherited_protocol_methods(
-        &self,
-        _protocol_id: TypeId,
-        parent_id: TypeId,
-    ) -> Vec<InheritedProtocolMethod> {
-        let mut out = Vec::new();
-        let mut cursor = Some(parent_id);
-        while let Some(current) = cursor {
-            let is_protocol = self
-                .analyzer
-                .type_table
-                .get_struct(current)
-                .is_some_and(|info| info.is_protocol);
-            if is_protocol {
-                if let Some(info) = self.analyzer.type_table.get_struct(current) {
-                    for (method_name, method_type_id) in &info.methods {
-                        if out.iter().any(|m: &InheritedProtocolMethod| &m.name == method_name) {
-                            continue;
-                        }
-                        let key = SymbolCollector::method_symbol_key(current, method_name);
-                        if let Some(signature) = self.analyzer.functions.get(&key) {
-                            out.push(InheritedProtocolMethod {
-                                name: method_name.clone(),
-                                type_id: *method_type_id,
-                                param_types: signature.param_types.clone(),
-                                return_type: signature.return_type,
-                            });
-                        }
-                    }
-                }
-            }
-            cursor = self
-                .analyzer
-                .type_table
-                .get_struct(current)
-                .and_then(|info| info.parent);
-        }
-        out
-    }
-
     pub(super) fn validate_protocol_conformance(
         &self,
         impl_type: SemanticType,
         protocol_id: TypeId,
         source: &str,
     ) -> Option<()> {
-        let SemanticType::Struct(impl_raw) = impl_type else {
-            return None;
-        };
-        let impl_id = TypeId(impl_raw);
-        if self
-            .analyzer
-            .type_table
-            .get_struct(impl_id)
-            .is_some_and(|info| info.is_protocol)
-        {
-            return None;
-        }
-
-        let protocol_methods = self.collect_inherited_protocol_methods(protocol_id, protocol_id);
-        for method in &protocol_methods {
-            let impl_signature = self
-                .resolve_method_symbol_key(impl_id, &method.name)
-                .or_else(|| self.resolve_method_symbol_key_in_structs(impl_id, &method.name))
-                .and_then(|key| self.analyzer.functions.get(&key).cloned());
-
-            let Some(impl_signature) = impl_signature else {
-                return None;
-            };
-
-            if impl_signature.param_types.len() != method.param_types.len() {
-                return None;
-            }
-
-            for (impl_t, proto_t) in impl_signature
-                .param_types
-                .iter()
-                .zip(method.param_types.iter())
-            {
-                if !self.variance_param_compatible(*impl_t, *proto_t) {
-                    let _ = source;
-                    return None;
-                }
-            }
-
-            if !self.variance_return_compatible(impl_signature.return_type, method.return_type) {
-                return None;
-            }
-        }
-
-        Some(())
-    }
-
-    fn resolve_method_symbol_key_in_structs(
-        &self,
-        receiver: TypeId,
-        method_name: &str,
-    ) -> Option<String> {
-        let mut cursor = Some(receiver);
-        while let Some(current) = cursor {
-            if let Some(info) = self.analyzer.type_table.get_struct(current) {
-                if info.methods.iter().any(|(name, _)| name == method_name) {
-                    let key = SymbolCollector::method_symbol_key(current, method_name);
-                    if self.analyzer.function_symbols.contains_key(&key) {
-                        return Some(key);
-                    }
-                }
-                cursor = info.parent;
-            } else {
-                return None;
-            }
-        }
-        None
+        ProtocolChecker::validate_protocol_conformance(self.analyzer, impl_type, protocol_id, source)
     }
 
     pub(super) fn validate_protocol_method_call(
@@ -471,170 +361,7 @@ impl<'a> TypeChecker<'a> {
         call_span: Span,
         source: &str,
     ) -> Option<SemanticType> {
-        let SemanticType::Struct(impl_id_raw) = impl_type else {
-            return None;
-        };
-        let impl_id = TypeId(impl_id_raw);
-
-        if self
-            .analyzer
-            .type_table
-            .get_struct(impl_id)
-            .is_some_and(|info| info.is_protocol)
-        {
-            return None;
-        }
-
-        let impl_method_key = self.resolve_method_symbol_key(impl_id, method_name);
-        let impl_signature = impl_method_key
-            .as_ref()
-            .and_then(|key| self.analyzer.functions.get(key).cloned());
-
-        let protocol_methods = self.collect_inherited_protocol_methods(protocol_id, protocol_id);
-        let Some(protocol_signature) = protocol_methods
-            .iter()
-            .find(|m| m.name == method_name)
-            .map(|m| FunctionSignature {
-                type_id: m.type_id.0,
-                param_types: m.param_types.clone(),
-                return_type: m.return_type,
-            })
-        else {
-            self.analyzer.push_semantic_error(
-                call_span,
-                source,
-                format!(
-                    "Type does not conform to protocol: method '{}' is not declared in the protocol.",
-                    method_name
-                ),
-            );
-            return None;
-        };
-
-        let Some(impl_signature) = impl_signature else {
-            self.analyzer.push_semantic_error(
-                call_span,
-                source,
-                format!(
-                    "Type does not conform to protocol: method '{}' is not declared in the implementing type.",
-                    method_name
-                ),
-            );
-            return None;
-        };
-
-        if impl_signature.param_types.len() != protocol_signature.param_types.len() {
-            self.analyzer.push_type_error(
-                call_span,
-                source,
-                format!(
-                    "Type does not conform to protocol: method '{}' expects {} argument(s), but type provides {}.",
-                    method_name,
-                    protocol_signature.param_types.len(),
-                    impl_signature.param_types.len()
-                ),
-            );
-            return None;
-        }
-
-        for (index, (impl_t, proto_t)) in impl_signature
-            .param_types
-            .iter()
-            .zip(protocol_signature.param_types.iter())
-            .enumerate()
-        {
-            if !self.variance_param_compatible(*impl_t, *proto_t) {
-                self.analyzer.push_type_error(
-                    call_span,
-                    source,
-                    format!(
-                        "Type does not conform to protocol: method '{}' argument #{} has incompatible variance: expected {} (contravariant) in protocol, got {}.",
-                        method_name,
-                        index + 1,
-                        proto_t.display_name(),
-                        impl_t.display_name()
-                    ),
-                );
-                return None;
-            }
-        }
-
-        if !self.variance_return_compatible(impl_signature.return_type, protocol_signature.return_type)
-        {
-            self.analyzer.push_type_error(
-                call_span,
-                source,
-                format!(
-                    "Type does not conform to protocol: method '{}' return type is incompatible (covariant): expected {} in protocol, got {}.",
-                    method_name,
-                    protocol_signature.return_type.display_name(),
-                    impl_signature.return_type.display_name()
-                ),
-            );
-            return None;
-        }
-
-        Some(impl_signature.return_type)
-    }
-
-    fn variance_param_compatible(
-        &self,
-        impl_type: SemanticType,
-        protocol_type: SemanticType,
-    ) -> bool {
-        if impl_type == SemanticType::Unknown || protocol_type == SemanticType::Unknown {
-            return true;
-        }
-        if impl_type == protocol_type {
-            return true;
-        }
-        let SemanticType::Struct(impl_raw) = impl_type else {
-            return false;
-        };
-        let SemanticType::Struct(proto_raw) = protocol_type else {
-            return false;
-        };
-        let impl_id = TypeId(impl_raw);
-        let proto_id = TypeId(proto_raw);
-        if self
-            .analyzer
-            .type_table
-            .get_struct(proto_id)
-            .is_some_and(|info| info.is_protocol)
-        {
-            return false;
-        }
-        self.is_subtype_of(proto_id, impl_id)
-    }
-
-    fn variance_return_compatible(
-        &self,
-        impl_type: SemanticType,
-        protocol_type: SemanticType,
-    ) -> bool {
-        if impl_type == SemanticType::Unknown || protocol_type == SemanticType::Unknown {
-            return true;
-        }
-        if impl_type == protocol_type {
-            return true;
-        }
-        let SemanticType::Struct(impl_raw) = impl_type else {
-            return false;
-        };
-        let SemanticType::Struct(proto_raw) = protocol_type else {
-            return false;
-        };
-        let impl_id = TypeId(impl_raw);
-        let proto_id = TypeId(proto_raw);
-        if self
-            .analyzer
-            .type_table
-            .get_struct(proto_id)
-            .is_some_and(|info| info.is_protocol)
-        {
-            return false;
-        }
-        self.is_subtype_of(impl_id, proto_id)
+        ProtocolChecker::validate_protocol_method_call(self.analyzer, impl_type, protocol_id, method_name, call_span, source)
     }
 
     pub(in crate::semantic) fn check_type_decl(&mut self, type_decl: &TypeDecl, source: &str) {
